requirements.txt
absl-py==1.3.0
arabic-reshaper==2.1.3
asgiref==3.5.2
astunparse==1.6.3
backports.entry-points-selectable==1.1.0
beartype==0.10.4
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==37.0.4
distlib==0.3.3
Django==4.1
django-admin-interface==0.18.2
django-colorfield==0.4.5
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-rest-knox==4.2.0
django-restframework==0.0.1
django-widget-tweaks==1.4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
env==0.1.0
filelock==3.3.2
Flask==2.2.2
flatbuffers==22.10.26
flet==0.1.41
future==0.18.2
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
gunicorn==20.1.0
h5py==3.7.0
html5lib==1.1
idna==3.4
img==2.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
MouseInfo==0.1.3
nltk==3.7
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pbr==5.9.0
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.6
pyasn==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.3.0
PyMsgBox==1.0.9
pyparsing==3.0.9
PyPDF2==1.26.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-bidi==0.4.2
python-memcached==1.59
pytweening==1.0.4
pytz==2022.2.1
qrcode==7.3.1
regex==2022.10.31
reportlab==3.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
stevedore==4.0.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.10.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
xhtml2pdf==0.2.5
